Пример новой реализации стримов и примеры
