+++ core/src/main/java/ru/tinkoff/piapi/core/stream/StreamObserverWithProcessor.java
@@ -1,0 +1,33 @@
+package ru.tinkoff.piapi.core.stream;
+
+import io.grpc.stub.StreamObserver;
+
+import java.util.function.Consumer;
+
+public class StreamObserverWithProcessor<T> implements StreamObserver<T> {
+
+  private final StreamProcessor<T> streamProcessor;
+  private final Consumer<Throwable> onErrorCallback;
+
+  public StreamObserverWithProcessor(StreamProcessor<T> streamProcessor,
+                                     Consumer<Throwable> onErrorCallback) {
+    this.streamProcessor = streamProcessor;
+    this.onErrorCallback = onErrorCallback;
+  }
+
+  @Override
+  public void onNext(T value) {
+    streamProcessor.process(value);
+  }
+
+  @Override
+  public void onError(Throwable t) {
+    if (onErrorCallback != null) {
+      onErrorCallback.accept(t);
+    }
+  }
+
+  @Override
+  public void onCompleted() {
+  }
+}
